add GameAdminActions
Add a new GameAdminActions component that is used by all the
game history pages to provide a common set of admin action buttons
of "reload" and "import game"
