+++ client/src/games/components/PastGamesCalendarPage.test.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { screen } from '@testing-library/react';
 import log from 'loglevel';
 
 import { fetchMock, renderWithProviders } from '../../../tests';
 import { PastGamesCalendarPage } from './PastGamesCalendarPage';
-import { MockBingoServer, normalUser } from '../../../tests/MockServer';
+import { MockBingoServer, adminUser } from '../../../tests/MockServer';
 import { createStore, initialState } from '../../store';
 
 
@@ -15,14 +14,18 @@ describe('PastGamesCalendarPage component', () => {
     });
 
     it('to render a calendar previous games', async () => {
-        //log.setLevel('debug');
-        const mockServer = new MockBingoServer(fetchMock, { currentUser: normalUser });
+        const mockServer = new MockBingoServer(fetchMock, { currentUser: adminUser });
+        const user = mockServer.getUserState(adminUser);
         const store = createStore({
             ...initialState,
-            user: mockServer.getUserState(normalUser),
+            games: {
+                ...initialState.games,
+                user: user.pk,
+            },
+            user,
         });
-        const { asFragment } = renderWithProviders(<PastGamesCalendarPage />, { store });
-        await screen.findAllByText("Rock & Power Ballads", { exact: false });
+        const { asFragment, findAllByText } = renderWithProviders(<PastGamesCalendarPage />, { store });
+        await findAllByText("Rock & Power Ballads", { exact: false });
         await fetchMock.flush(true);
         expect(asFragment()).toMatchSnapshot();
     });
